fix(lint): split routine_storage[_tests].rs under the 600-line gate
routine_storage.rs (605) and routine_storage_tests.rs (661) had crept past
the repo's 600-line pre-push gate. Extracts the one-time startup migration
functions into routine_storage_migrations.rs, and the runtime-sidecar
(manual-trigger/power-saving) tests into routine_storage_sidecar_state_tests.rs.
No behavior or test-body changes.
